Rename component file ButtonRedux -> SaveButton
